+++ default.nix
@@ -1,1 +1,9 @@
-(import ./release.nix {}).flay
+{ nixpkgs ? import ./nixpkgs.nix }:
+
+let
+pkgs = import nixpkgs {};
+ghc841 = pkgs.haskell.packages.ghc841.override {
+  packageSetConfig = import ./hs-overlay.nix;
+};
+
+in { inherit (ghc841) flay; }